backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor==3.3.2
pymongo==4.6.2
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.31.0
pdfplumber==0.10.3
aiofiles==23.2.1
httpx==0.27.0
pytesseract==0.3.13
